fix: keep slot outlet v-on ci gates green
