+++ tests/tooling/davinci-v-on-storage.test.ts
@@ -170,7 +170,7 @@ test("the natural committed v-on corpus fits the two-entry inline buckets", () =
     { options: 0, event: 0, keys: 0 },
   );
 
-  assert.equal(spellings.length, 188, "update the measured corpus evidence intentionally");
+  assert.equal(spellings.length, 194, "update the measured corpus evidence intentionally");
   assert.deepEqual(maxima, { options: 2, event: 2, keys: 2 });
 });
